adding buttons for additions to simulator, ability to add to request queue
